+++ src/main/java/net/sf/jabref/gui/menus/RightClickMenu.java
@@ -67,6 +67,9 @@ public class RightClickMenu extends JPopupMenu implements PopupMenuListener {
         copySpecialMenu.add(new GeneralAction(Actions.COPY_CITE_KEY, Localization.lang("Copy \\cite{BibTeX key}"), KeyBinding.COPY_CITE_BIBTEX_KEY));
         copySpecialMenu.add(new GeneralAction(Actions.COPY_KEY_AND_TITLE, Localization.lang("Copy BibTeX key and title"), KeyBinding.COPY_BIBTEX_KEY_AND_TITLE));
         copySpecialMenu.add(new GeneralAction(Actions.COPY_KEY_AND_LINK, Localization.lang("Copy BibTeX key and link"), KeyBinding.COPY_BIBTEX_KEY_AND_LINK));
+        copySpecialMenu.add(new GeneralAction(Actions.COPY_PREVIEW_OF_SELECTED_ENTRIES_TO_CLIPBOARD,
+                Localization.lang("Copy preview of selected entries"),
+                KeyBinding.COPY_PREVIEW_OF_SELECTED_ENTRIES_TO_CLIPBOARD));
         copySpecialMenu.add(new GeneralAction(Actions.EXPORT_TO_CLIPBOARD, Localization.lang("Export to clipboard"),
                 IconTheme.JabRefIcon.EXPORT_TO_CLIPBOARD.getSmallIcon()));